requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
assemblyai==0.17.0
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==7.1.2
dataclasses-json==0.5.14
Flask==1.1.4
Flask-Sockets==0.2.1
frozenlist==1.4.0
gevent==23.9.0.post1
gevent-websocket==0.10.1
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-speech==2.21.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.3
langchain==0.0.279
langsmith==0.0.33
MarkupSafe==2.0.1
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.28.0
packaging==23.1
proto-plus==1.22.3
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
PyJWT==2.8.0
python-vlc==3.0.18122
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
SpeechRecognition==3.10.0
SQLAlchemy==2.0.20
tenacity==8.2.3
tqdm==4.66.1
twilio==8.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==1.0.1
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
